# src/docs/swaggerdoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/shifts": {
      "post": {
        "summary": "Create a new work shift",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "startTime", "endTime"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Morning Shift"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "time",
                    "example": "08:00"
                  },
                  "endTime": {
                    "type": "string",
                    "format": "time",
                    "example": "16:00"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Work shift created successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "get": {
        "summary": "Get all work shifts",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of work shifts"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/shifts/{shiftId}": {
      "get": {
        "summary": "Get work shift by ID",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shiftId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Work shift details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Work shift not found"
          }
        }
      },
      "patch": {
        "summary": "Update work shift",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shiftId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "startTime": {
                    "type": "string",
                    "format": "time"
                  },
                  "endTime": {
                    "type": "string",
                    "format": "time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Work shift updated successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Work shift not found"
          }
        }
      },
      "delete": {
        "summary": "Delete work shift",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "shiftId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Work shift deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Work shift not found"
          }
        }
      }
    },
    "/shifts/sessions/open": {
      "post": {
        "summary": "Open a new shift session",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["shiftId", "employeeId"],
                "properties": {
                  "shiftId": {
                    "type": "integer"
                  },
                  "employeeId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shift session opened successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/shifts/sessions/{sessionId}/close": {
      "post": {
        "summary": "Close a shift session",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shift session closed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/shifts/sessions/{sessionId}/approve": {
      "post": {
        "summary": "Approve a shift session",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Shift session approved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/shifts/sessions": {
      "get": {
        "summary": "Get all shift sessions",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of shift sessions"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/shifts/sessions/me": {
      "get": {
        "summary": "Get current user's shift session",
        "tags": ["Shifts"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current shift session details"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}